backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
ollama
# Stable LangChain 0.2 Stack
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-text-splitters==0.2.2
langchain-google-genai==1.0.10
google-generativeai==0.7.2
langchain-huggingface==0.0.3
huggingface-hub==0.23.0
sentence-transformers==3.0.0
# Vector DB
chromadb==0.5.0
langchain-chroma==0.1.2
